refactor/ separate copyRepo logic from copyStore
